+++ src/utils/jm_domain_detect.py
@@ -189,7 +189,12 @@ def move_jm_keyword_dir(actor, a_title_list):
     root_path_jm = os.path.join(constants.data_path, "jm_image")
     for actor_jm_content in a_title_list:
         try:
-            move_target_folder(os.path.join(root_path_jm, actor_jm_content), os.path.join(actor_path, actor_jm_content))
+            src_path = os.path.join(root_path_jm, actor_jm_content)
+            target_path =  os.path.join(actor_path,actor_jm_content)
+            if not move_target_folder(src_path, target_path):
+                continue
+            else:
+                logger.success(f"Move name: {actor_jm_content} Success, move to {target_path}!")
         except Exception as e:
             logger.warning(f"Unknown error, detail: {e}, skip: {actor_jm_content}.")
             continue
@@ -242,12 +247,13 @@ def search_download_jm(actor):
     for aid_process in aid_list:
         if search_local_jm_keyword(a_title_list[download_jm_index]):
             logger.warning(f"Keyword JM image title: {a_title_list[download_jm_index]} exists, skip.")
+            download_jm_index += 1
             continue
         else:
             download_album(aid_process, jm_option)
             logger.debug(f"Download jm image: {aid_process}, title: {a_title_list[download_jm_index]},"
                          f" download index: {download_jm_index + 1}, count: {len(aid_list)}, finish.")
-        download_jm_index += 1
+            download_jm_index += 1
         if not constants.JM_SD_auto_flag:
             return False
     # logger.success(f"Download JM keyword: {actor} image finish.")
@@ -309,7 +315,7 @@ def search_local_jm_keyword(jm_title):
     jm_root_dir = os.path.join(constants.data_path, "jm_image")
     for root, dirs, files in os.walk(jm_root_dir):
         for dir_name in dirs:
-            if jm_title.lower() in dir_name.lower():
+            if jm_title in dir_name:
                 return True
     return False