+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - add download finish keyword folder move
 - update move method
 - skip exists jm title
+- fix exists jm title bug
 
 ## v1.1.8-beta.1.240624
 ### 2024/6/24